+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `Header` element
 - Add `Button` element - primary, secondary and default styles
+- Add `SocialItem` element
+  - Supported socials:
+    - "LinkedIn", "Facebook", "Twitter", "Reddit", "GitHub", "Youtube",
+      "Pinterest", "Instagram", "Tumbler", "Medium", "TikTok"
 
 ## [0.0.1] - 2021-02-20